requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
Deprecated==1.2.10
Django==3.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
dnspython==2.0.0
emoji==0.5.4
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pydig==0.3.0
PyGithub==1.51
PyJWT==1.7.1
pylint==2.5.3
pytz==2020.1
requests==2.24.0
semver==2.9.1
six==1.12.0
sqlparse==0.3.1
timeout-decorator==0.4.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
django-environ==0.4.5
